Changing parsing spec from manually to pyrpm tool
- pyrpm tool seems to be a little bit abandoned, but for now it's
ok. Later I'll decide if keep using it or going for a different one.
- A current issue with the stx graph generation is that many nodes
are generated for the same deb package. This happens because in this
script a node is generated for every BuildRequires.
The very next problem to fix is change the code to use an existing
node instead of creating a new one. And then doing the add_edge on
that existing node.

Signed-off-by: Marcela Rosales <marcelarosalesj@gmail.com>